pyproject.toml
[project]
name = "dripdrop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "boto3>=1.38.38",
    "bs4>=0.0.2",
    "celery[redis]>=5.5.3",
    "commitizen>=4.8.3",
    "fake-useragent>=2.2.0",
    "faker>=37.4.0",
    "fastapi[all]>=0.115.13",
    "httpx>=0.28.1",
    "mutagen>=1.47.0",
    "pre-commit>=4.2.0",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "smtp2go>=2.3.1",
    "sqlalchemy>=2.0.41",
    "yt-dlp>=2025.6.9",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true

[tool.pytest.ini_options]
addopts = "--strict-markers"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
markers = [
    "long: marks test as slow"
]

[tool.coverage.run]
omit = [
    "*/db/migrations/*"
]
